feat(transactions): empty state for transaction list
The heading and actions (filters and download button) were moved from the transaction-list component to the transactions-summary-page for better component separation and reusability. The transaction-list component was simplified to focus solely on rendering the list of transactions. Additionally, an empty state with a new icon was added to handle cases where no transactions are found.